backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
anthropic==0.18.0
openai==1.12.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
python-multipart==0.0.6
httpx==0.26.0
aiofiles==23.2.1
python-docx==1.1.0
reportlab==4.0.9
markdown==3.5.2

